Created a link to the python backend
I created a PythonCaller class which calls the python file at the directory you specify. The python file needs to PRINT the result and then the result will be read and returned. Currently only works with strings. Complex data types will need to be serialised into formats such as json.

The File python caller has the function call in it which is used to call a python script.

The file webcontroller has been modified for the purpose of demonstration. Visiting /python will call this demonstration.

The file test.py was created as a dummy for the python backend.