Arguments
Added the ability for the python caller to take arguments. This will allow for specific functions to be executed in the python scripts and for the scripts to take arguments. These arguments may be values in the database, a pointer to a value in the database, requirements to be iterated upon, or more. One caveat is that EVERYTHING must be passed as a string and then converted to the correct type. I am looking at ways to avoid this but I am not sure if it is possible.