Python Unit Testing
renamed test.py to testModule.py and added a new file called test.py.
testModule is a module to simulate one of the tools the NCC uses in its toolchain, specifically for optimising components.
test.py is used to run unit tests on the testModule.
As stated in test.py, a seperate folder to keep the tests in in preferable but causes a lot of errors. Details in the comment in test.py.