Continued work on new CSS
Projects now appear as buttons, a maximum of 4 buttons per row and then they start a new row. The pop up to add a new project no longer moves down the screen with the project buttons but instead always appears in the middle of the user's screen. Implementation of a cancel new project addition button/link using some javascript